Fix a validador de transparencia.

+++ src/main/java/domain/entities/operaciones/Item.java
@@ -59,7 +59,7 @@ public class Item extends EntidadPersistente {
         return tipo;
     }
 
-    public float getCantidad() {
+    public int getCantidad() {
         return cantidad;
     }
